requirements.txt
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.3.0
langchain-text-splitters>=1.1.0
langchain-groq>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
